tests/wpt/web-platform-tests/webstorage/storage_session_setitem_quotaexceedederr.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <meta name="timeout" content="long">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>storage_session_setItem_QUOTA_EXCEEDED_ERR</h1>
    <div id="log"></div>
    <script>
        test(function() {
            sessionStorage.clear();

            var index = 0;
            var key = "name";
            var val = "user";

            assert_throws("QUOTA_EXCEEDED_ERR", function() {
                while (true) {
                    index++;
                    sessionStorage.setItem("" + key + index, "" + name + index);
                }
            });

            sessionStorage.clear();
        });
    </script>
 </body>
</html>
